move amazonka-related definitions to Amazonka, clean some imports
