+++ src/Upcast/Infra/Resources/Subnet.hs
@@ -4,10 +4,10 @@
 
 module Upcast.Infra.Resources.Subnet where
 
-import Control.Applicative
+import Control.Applicative -- (*)
 import Control.Monad.Trans (liftIO)
 import Control.Lens hiding (Context) -- (*)
-import Control.Lens.Action
+import Control.Lens.Action -- (*)
 import Control.Monad.Trans.AWS (send)
 import Control.Monad.Reader (MonadReader, asks)
 import Control.Monad (void)
@@ -19,9 +19,9 @@ import Data.Text (Text, unpack)
 
 import qualified Network.AWS.EC2 as EC2 -- (*)
 
-import Upcast.Infra.Types.Amazonka (ResourceId)
+import Upcast.Infra.Types.Amazonka -- (*)
 import Upcast.Infra.NixTypes -- (*)
-import Upcast.Infra.Types.Common
+import Upcast.Infra.Types.Common -- (*)
 
 -- *
